+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [next version] - 2020-09-14
+- Updated Instruments and routines to conform with changes made for pysat 3.0
 - New Instruments
   - GOLD Nmax